rename a few things
